Change time to float in the results

+++ Trivia Project/Trivia Project/GameManager.h
@@ -14,7 +14,7 @@ struct PlayerResults
 	std::string username;
 	unsigned int correctAnswersCount;
 	unsigned int wrongAnswersCount;
-	unsigned int averageAnswerTime;
+	float averageAnswerTime;
 
 	std::string toString()
 	{
